Cleaned up k8sclient package
